MapCollision::is_wall: change signature to accept float
Nearly all the other functions in MapCollision accept float, so this
must have been overlooked. By changing the function to accept float, we
can remove the rounding/flooring from all calling points moving it into
the function itself.

Similar as is_empty(float, float) this also needs to access the the
collision layer, so it needs to transform the floats to integers.
In is_empty we chose floor, which makes sense here as well. This is a
change of semantics (before some places used round instead of floor).

+++ src/MapCollision.h
@@ -78,7 +78,7 @@ public:
 
 	bool is_outside_map(float tile_x, float tile_y) const;
 	bool is_empty(float x, float y) const;
-	bool is_wall(int x, int y) const;
+	bool is_wall(float x, float y) const;
 
 	bool is_valid_position(float x, float y, MOVEMENTTYPE movement_type, bool is_hero) const;
